+++ packages/common/src/consts/index.ts
@@ -174,6 +174,37 @@ export const MatchReasonCode = {
 export const URL_PATTERN =
   /^(?:https?:\/\/)?(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z0-9\u00A1-\uFFFF][a-z0-9\u00A1-\uFFFF_-]{0,62})?[a-z0-9\u00A1-\uFFFF]\.)+(?:[a-z\u00A1-\uFFFF]{2,}\.?))(?::\d{2,5})?(?:[/?#]\S*)?$/i;
 
+export const CHECKS_STATUSES = ['in-progress', 'failed', 'completed'] as const;
+
+export type ChecksStatus = (typeof CHECKS_STATUSES)[number];
+
+export const CHECKS_STATUSES_MAP = Object.fromEntries(
+  CHECKS_STATUSES.map(status => [status, status]),
+) as { [K in ChecksStatus]: K };
+
+export const ASSESSMENT_STATUSES = [
+  'in-progress',
+  'pending-review',
+  'under-review',
+  'approved',
+  'rejected',
+] as const;
+
+export type AssessmentStatus = (typeof ASSESSMENT_STATUSES)[number];
+
+export const ASSESSMENT_STATUSES_MAP = Object.fromEntries(
+  ASSESSMENT_STATUSES.map(status => [status, status]),
+) as { [K in AssessmentStatus]: K };
+
+export const UPDATEABLE_ASSESSMENT_STATUSES = [
+  ASSESSMENT_STATUSES_MAP['pending-review'],
+  ASSESSMENT_STATUSES_MAP['under-review'],
+  ASSESSMENT_STATUSES_MAP['approved'],
+  ASSESSMENT_STATUSES_MAP['rejected'],
+] as const;
+
+export type UpdateableAssessmentStatus = (typeof UPDATEABLE_ASSESSMENT_STATUSES)[number];
+
 export const MERCHANT_REPORT_STATUSES = [
   'in-progress',
   'quality-control',
@@ -266,3 +297,5 @@ Object.entries(SUPPORTED_FILE_EXT_ENUM).forEach(([key, value]) => {
 
 export type SupportedFileExt =
   (typeof SUPPORTED_FILE_EXT_ENUM)[keyof typeof SUPPORTED_FILE_EXT_ENUM];
+
+export const BANK_ACCOUNT_VERIFICATION_COMMERCIAL_REQUEST_TYPE = 'BAVCommercial-Standard' as const;